#5867 simplifying test case by inlining all required models into the test case
